+++ lib/src/image_searcher.dart
@@ -17,8 +17,13 @@ class ImageSearcher {
   /// The SauceNAO user info such as account type and account search limit
   UserInfo user = UserInfo();
 
+  Function(UserInfo userInfo)? onUserInfoGot;
+
   /// init
-  ImageSearcher({required this.searcherConfig, Uri? uri});
+  ImageSearcher({
+    required this.searcherConfig,
+    Uri? uri,
+  });
 
   /// Search Image from Uri
   Future<SearchResult?> uri(
@@ -56,8 +61,8 @@ class ImageSearcher {
     Map infoMap = jsonDecode(resJson);
 
     // If meets Api Exception
-    if (infoMap['header']['status'] == -1) {
-      throw ApiStatusException(infoMap: infoMap);
+    if (infoMap['header']['status'] != 0) {
+      throw genApiErrorFromInfoMap(infoMap);
     }
 
     // update SauceNAO user info
@@ -99,74 +104,13 @@ class ImageSearcher {
     }
     Map infoMap = jsonDecode(jsonStr);
 
+    // update user info
+    user.update(infoMap);
+    // callback Function (if have)
+    if (onUserInfoGot != null) {
+      onUserInfoGot!(user);
+    }
+
     return SearchResult.fromMap(infoMap);
   }
 }
-
-// -----------------------------------------------------------
-// UserInfo and LimitInfo
-
-/// SauceNAO UserInfo class, contains user type, user id,
-/// and user API limit info
-class UserInfo {
-  /// The SauceNAO Id of the user
-  int? id;
-  // The user type of the user, check SauceNAO website for more info
-  int? type;
-
-  /// The total successful request count of this account
-  int? requested;
-
-  /// The limit info of this user
-  LimitInfo limit = LimitInfo();
-
-  void update(Map infoMap) {
-    try {
-      // If API Exception
-      if (infoMap['header']['status'] == -1) {
-        return;
-      }
-
-      // id
-      try {
-        id = int.parse(infoMap['header']['user_id']);
-      } catch (e) {}
-
-      // type
-      try {
-        id = int.parse(infoMap['header']['account_type']);
-      } catch (e) {}
-
-      // limit
-      limit = LimitInfo.fromInfoMap(infoMap);
-    } catch (e) {}
-  }
-}
-
-/// SauceNAO user limit info. Contains Long/Short limit info.
-///
-/// About the limit of SauceNAO API, please check
-/// [SauceNAO Official Website](https://saucenao.com)
-/// for more info
-class LimitInfo {
-  int? short;
-  int? long;
-  int? shortRemaining;
-  int? longRemaining;
-
-  LimitInfo() {
-    ;
-  }
-
-  /// Create user limit info from `Map` type object
-  LimitInfo.fromInfoMap(infoMap) {
-    try {
-      short = int.parse(infoMap['header']['short_limit']);
-      long = int.parse(infoMap['header']['long_limit']);
-      shortRemaining = int.parse(infoMap['header']['short_remaining']);
-      longRemaining = int.parse(infoMap['header']['long_remaining']);
-    } catch (e) {}
-  }
-}
-
-// ---------------------------------------------------