+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.1
+- Update: Add new Exception type: `ApiRateLimitExceededException`
+
+## 0.2.0
+- Update: Add new Exception type: `ApiLimitExceededException`
+- Update: Allow to add callback function to deal with userInfo.
+- Fix: The userInfo in ImageSearcher never update
+
+## 0.1.4
+- Update: Add default title to `SearchResultItem`
+
 ## 0.1.3
 - Update `README.md`
 - Fix the issue that the Exceptions is not visible